Update HTML and CSS
The overall design and layout is done. The links go to where they should, the items are centered the way they should, and the Photography section wraps. The site looks good in my web broswer, but does not look good on my phone. The CSS needs to be updated to fit to mobile device as well. As  of right now, the image breaks its container when the screen is re-sized too small. Once this is fixed, add a README w/ screenshots of the site.